move page spec test files
